VPD-1472: Move specs and add for GroupId
